cloud/requirements.txt
# Cloud backend dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0
pydantic-settings>=2.0

# Database
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20.0        # SQLite async (dev)
# asyncpg>=0.30.0        # PostgreSQL async (prod) — uncomment for production

# Auth — Supabase JWT verification
PyJWT>=2.8.0

# Testing
pytest>=8.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
